Revert "Fix defining types conditionally and/or in subshells (re: 1dc18346)"
This reverts commit 5d3532100679911fa980bf8909a3539e8a826945.

This is not ready for the main development branch as it's still
broken, distracting from other issues that also need fixing.

It crashses when running a $PS2 discipline function as the
recursion tracking gets messed up. It is another context where
arbitrary code is executed at parse time!

I'm putting this one on the back burner. So that terrible problem
with types is back, sorry about that.